✨ feat: 修复 arm64 交叉编译的 apt 源配置 / Fix apt sources for arm64 cross-compile

+++ src-tauri/src/commands/settings_cmd.rs
@@ -249,9 +249,12 @@ pub fn get_disk_space_inner(output_dir: &str) -> Result<DiskSpace> {
         let ret = unsafe { libc::statvfs(path_cstr.as_ptr(), stat.as_mut_ptr()) };
         if ret == 0 {
             let stat = unsafe { stat.assume_init() };
-            let block = stat.f_frsize;
-            let total = stat.f_blocks * block;
-            let avail = stat.f_bavail * block;
+            #[allow(clippy::unnecessary_cast)]
+            let block = stat.f_frsize as u64;
+            #[allow(clippy::unnecessary_cast)]
+            let total = stat.f_blocks as u64 * block;
+            #[allow(clippy::unnecessary_cast)]
+            let avail = stat.f_bavail as u64 * block;
             return Ok(DiskSpace {
                 total_bytes: total,
                 available_bytes: avail,
